+++ assets/js/annotation.js
@@ -142,7 +142,7 @@ class Annotation extends Track {
 
       const textYPos = this.tracksYPos(heightOrder);
       // limit drawing of titles to certain resolution
-      if ( this.getResolution < 5) {
+      if ( this.getResolution < 6) {
         // Draw annotation name
         this.heightOrderRecord.latestNameEnd = this.drawText(
           annotationName,
@@ -168,14 +168,14 @@ chr${chrom}:${start}-${end}
 Score = ${score}`;
 
       // Add tooltip title for whole gene
-      this.heightOrderRecord.latestTrackEnd = this.hoverText(
-        geneText,
-        `${scale * (start - this.trackData.start_pos)}px`,
-        `${textYPos - this.featureHeight / 2}px`,
-        `${scale * (end - start)}px`,
-        `${this.featureHeight + textSize}px`,
-        0,
-        this.heightOrderRecord.latestHeight);
+      // this.heightOrderRecord.latestTrackEnd = this.hoverText(
+      //   geneText,
+      //   `${scale * (start - this.trackData.start_pos)}px`,
+      //   `${textYPos - this.featureHeight / 2}px`,
+      //   `${scale * (end - start)}px`,
+      //   `${this.featureHeight + textSize}px`,
+      //   0,
+      //   this.heightOrderRecord.latestHeight);
     }
   }
 }