+++ assets/js/track.js
@@ -1,5 +1,31 @@
 // Generic functions related to annotation tracks
 
+// function for shading and blending colors on the fly
+const pSBC=(p,c0,c1,l)=>{
+    let r,g,b,P,f,t,h,i=parseInt,m=Math.round,a=typeof(c1)=="string";
+    if(typeof(p)!="number"||p<-1||p>1||typeof(c0)!="string"||(c0[0]!='r'&&c0[0]!='#')||(c1&&!a))return null;
+    if(!this.pSBCr)this.pSBCr=(d)=>{
+        let n=d.length,x={};
+        if(n>9){
+            [r,g,b,a]=d=d.split(","),n=d.length;
+            if(n<3||n>4)return null;
+            x.r=i(r[3]=="a"?r.slice(5):r.slice(4)),x.g=i(g),x.b=i(b),x.a=a?parseFloat(a):-1
+        }else{
+            if(n==8||n==6||n<4)return null;
+            if(n<6)d="#"+d[1]+d[1]+d[2]+d[2]+d[3]+d[3]+(n>4?d[4]+d[4]:"");
+            d=i(d.slice(1),16);
+            if(n==9||n==5)x.r=d>>24&255,x.g=d>>16&255,x.b=d>>8&255,x.a=m((d&255)/0.255)/1000;
+            else x.r=d>>16,x.g=d>>8&255,x.b=d&255,x.a=-1
+        }return x};
+    h=c0.length>9,h=a?c1.length>9?true:c1=="c"?!h:false:h,f=this.pSBCr(c0),P=p<0,t=c1&&c1!="c"?this.pSBCr(c1):P?{r:0,g:0,b:0,a:-1}:{r:255,g:255,b:255,a:-1},p=P?p*-1:p,P=1-p;
+    if(!f||!t)return null;
+    if(l)r=m(P*f.r+p*t.r),g=m(P*f.g+p*t.g),b=m(P*f.b+p*t.b);
+    else r=m((P*f.r**2+p*t.r**2)**0.5),g=m((P*f.g**2+p*t.g**2)**0.5),b=m((P*f.b**2+p*t.b**2)**0.5);
+    a=f.a,t=t.a,f=a>=0||t>=0,a=f?a<0?t:t<0?a:a*P+t*p:0;
+    if(h)return"rgb"+(f?"a(":"(")+r+","+g+","+b+(f?","+m(a*1000)/1000:"")+")";
+    else return"#"+(4294967296+r*16777216+g*65536+b*256+(f?m(a*255):0)).toString(16).slice(1,f?undefined:-2)
+}
+
 // Check if two geometries are overlapping
 // each input is an object with start/ end coordinates
 // f          >----------------<
@@ -229,15 +255,12 @@ class Track {
     this.drawCtx.strokeStyle = color;
     this.drawCtx.lineWidth = lineWidth;
     this.drawCtx.beginPath();
-    console.log(`Move pointer to: ${xStart}, ${yPos}`)
     this.drawCtx.moveTo(xStart, yPos);  // begin at bottom left
     const waveLength = 2 * (height / Math.tan(45));
     const lineLength = xStop - xStart + 1;
-    console.log(`Start pos: ${xStart}, ${yPos}; Height: ${height}; WaveLength: ${waveLength}; Line len: ${lineLength}`)
     // plot whole wave pattern
     let midline = yPos - height / 2 // middle of line
     let lastXpos = xStart;
-    console.log(`Plot ${Math.floor(lineLength / (waveLength / 2))} full wave pattens`)
     for (let i = 0; i < Math.floor(lineLength / (waveLength / 2)); i++){
       lastXpos += waveLength / 2;
       height *= -1  // reverse sign
@@ -249,7 +272,6 @@ class Track {
       height *= -1  // reverse sign
       let partialWaveHeight = partialWaveLength * Math.tan(45);
       this.drawCtx.lineTo(xStop, yPos - Math.sign(height) * partialWaveHeight);
-      console.log(`Plot partial wave: ${partialWaveHeight}; Height: ${height}`)
     }
     this.drawCtx.stroke();
     this.drawCtx.restore();
@@ -289,11 +311,11 @@ class Track {
   // Draw an arrow in desired direction
   // Forward arrow: direction = 1
   // Reverse arrow: direction = -1
-  async drawArrow (xpos, ypos, direction, height, color) {
+  async drawArrow (xpos, ypos, direction, height, lineWidth=2, color) {
     let width = direction * this.arrowWidth;
     this.drawCtx.save();
     this.drawCtx.strokeStyle = color;
-    this.drawCtx.lineWidth = this.arrowThickness;
+    this.drawCtx.lineWidth = lineWidth;
     this.drawCtx.beginPath();
     this.drawCtx.moveTo(xpos - width / 2, ypos - height / 2);
     this.drawCtx.lineTo(xpos + width / 2, ypos);