+++ CHANGELOG.md
@@ -9,7 +9,13 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
  - Added github workflow for running pytest on PRs
  - Added unit tests
 ### Changed
+ - Changed positive strand color of transcripts to a more constrasting color
+ - Temporarily disabled on hover popups in annotation tracks
+ - Transcripts are represented as arrows in lower resolutions
+ - Highlight MANE transcript in name and with a brighter color
 ### Fixed
+ - Gene names are now centered below transcript
+ - Fixed assignement of height order when updating transcript data
 
 ## [1.1.2]
 ### Added